Removed some old VC6 workarounds, and removed the VC6 exporter from the introjucer.

+++ modules/juce_graphics/fonts/juce_Font.cpp
@@ -81,7 +81,7 @@ public:
         }
 
         int replaceIndex = 0;
-        int bestLastUsageCount = std::numeric_limits<int>::max();
+        size_t bestLastUsageCount = std::numeric_limits<int>::max();
 
         for (i = faces.size(); --i >= 0;)
         {
